refactor: 2개의 NumberBaseball을 받아 Hint를 생성하는 것은 Hint의 책임
HintService가 Hint를 생성하도록 역할에 따라 기능을 분리함.
이에 따라 main에서 GameController 호출 시, HintItemService들은 HintService에게 필요함.
