+++ src/main/java/baseball/service/GameService.java
@@ -2,6 +2,7 @@ package baseball.service;
 
 import baseball.domain.NumberBaseball;
 import baseball.domain.hint.Hint;
+import baseball.service.hint.HintService;
 import baseball.view.EndView;
 import baseball.view.HintView;
 import baseball.view.InputView;
@@ -9,7 +10,8 @@ import baseball.view.RestartView;
 
 public interface GameService {
 
-    void playOneGame(InputView inputView, HintView hintView, EndView endView, RestartView restartView, NumberBaseball computerBaseball);
-
-    Hint createHint(NumberBaseball computerBaseball, NumberBaseball inputBaseball);
+    void playOneGame(InputView inputView, HintView hintView,
+                     EndView endView, RestartView restartView,
+                     HintService hintService,
+                     NumberBaseball computerBaseball);
 }
