+++ src/main/java/baseball/service/hint/HintService.java
@@ -1,0 +1,9 @@
+package baseball.service.hint;
+
+import baseball.domain.NumberBaseball;
+import baseball.domain.hint.Hint;
+
+public interface HintService {
+
+    Hint createHint(NumberBaseball computerBaseball, NumberBaseball inputBaseball);
+}
